// src/app/models/eventRegistrationModels.ts
export interface IEventModels 
  {
    eventName : string;
    description: string;
    starting_Time: string;
    ending_Time: string;
    objective: string;
    status: string;
    startDate: string;
    endDate: string;
    type: string;
    location: string;
    attachment: string;
    event_Image: string;
    sponsors: string;
    outcome: string;
    capacity: number;
  }